# out/server/src/main/java/org/openapitools/api/LoginApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-06-12T13:57:43.210224513Z[Etc/UTC]", comments = "Generator version: 7.7.0-SNAPSHOT")
@Validated
@Tag(name = "users", description = "Operations related to users")
public interface LoginApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /login : Validate login
     *
     * @param userName  (required)
     * @param userPassword  (required)
     * @return Login validation result (status code 200)
     */
    @Operation(
        operationId = "validateLogin",
        summary = "Validate login",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Login validation result", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Boolean.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/login",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Boolean> validateLogin(
        @NotNull @Parameter(name = "userName", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "userName", required = true) String userName,
        @NotNull @Parameter(name = "userPassword", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "userPassword", required = true) String userPassword
    ) {
        // return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
        System.out.println(userPassword);
        return new ResponseEntity<>(true, HttpStatus.OK);

    }

}